First upload to repository
The diaFRAM software is largely based on the Linny-R code I developed over the past decade.
The GUI is styled in different colors, but essentially is a stripped version of the Linny-R GUI. The same holds for the Virtual Machine.
This version 0.1.0 does not support simulation yet, and does not display links that span multiple levels in the activity hierarchy.