Improved "deep links"
Clicking on a "deep" link, i.e., a link arrow representing multiple connections, now prompts for selection of a specific connection.

It is now also prohibited to "descend" (by double-clicking) into a function that itself has one or more connections to other functions.
Likewise, drag-dropping a selection into such a function is prohibited.